+++ source/Messaging.Infrastructure/Configuration/CompositionRoot.cs
@@ -30,6 +30,7 @@ using Messaging.Application.Configuration;
 using Messaging.Application.Configuration.Authentication;
 using Messaging.Application.Configuration.Commands;
 using Messaging.Application.Configuration.DataAccess;
+using Messaging.Application.Configuration.TimeEvents;
 using Messaging.Application.IncomingMessages.RequestChangeOfSupplier;
 using Messaging.Application.OutgoingMessages;
 using Messaging.Application.OutgoingMessages.Common;
@@ -55,7 +56,6 @@ using Messaging.Infrastructure.Configuration.DataAccess;
 using Messaging.Infrastructure.Configuration.InternalCommands;
 using Messaging.Infrastructure.Configuration.Processing;
 using Messaging.Infrastructure.Configuration.Serialization;
-using Messaging.Infrastructure.Configuration.SystemTime;
 using Messaging.Infrastructure.IncomingMessages;
 using Messaging.Infrastructure.MarketEvaluationPoints;
 using Messaging.Infrastructure.MasterData.MarketEvaluationPoints;
