+++ source/Messaging.Infrastructure/Configuration/InternalCommands/ProcessInternalCommandsOnTimeHasPassed.cs
@@ -16,7 +16,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
-using Messaging.Infrastructure.Configuration.SystemTime;
+using Messaging.Application.Configuration.TimeEvents;
 
 namespace Messaging.Infrastructure.Configuration.InternalCommands
 {
